requirements.txt
opencv-python==4.1.0.25
phonenumbers==8.10.15
Pillow==6.1.0
pytesseract==0.2.7
selenium==3.141.0
